server/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
python-jose[cryptography]
passlib[bcrypt]
bcrypt>=4.0.1
python-multipart
requests
python-dotenv
